Page Reload: Add case for save-as.
When user saves the document with a different name, new document will be
loaded.

While this happens, if there are more than one servers, client may be
directed to another server for the new file.

In this case, socket code will think that current server has restarted
and will reload the page. Socket code can't see if the server is
different than the previos one.

So, after a save-as operation, if we check the file adress with the
previous one, we can detect that the file we are viewving has changed.
In that case, we will not reload the page.

Change-Id: Idbf50678a453feccb34eb67a5b9286904a96bffa